register controller

+++ src/utils/cloudinary.js
@@ -1,30 +1,46 @@
 import { v2 as cloudinary } from "cloudinary";
 import fs from "fs";
 
+
 cloudinary.config({
   cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
   api_key: process.env.CLOUDINARY_API_KEY,
   api_secret: process.env.CLOUDINARY_API_SECRET,
 });
 
+console.log("ENV CHECK:", {
+  cloud: process.env.CLOUDINARY_CLOUD_NAME,
+  key: process.env.CLOUDINARY_API_KEY,
+  secret: process.env.CLOUDINARY_API_SECRET ? "LOADED" : "MISSING",
+} );
+
 const uploadOnCloudinary = async (localFilePath) => {
   try {
     if (!localFilePath) return null;
-    // upload the file on cloudinary
+    
+
     const response = await cloudinary.uploader.upload(localFilePath, {
-      resource_type: "auto",
+      resource_type: "image",
     });
-    // file has been uploaded successfully
-    console.log("File is uploaded on cloudinary", response.url);
+
+    // console.log(" Uploaded to Cloudinary:", response.secure_url);
+
+    //  safely delete temp file
+    if (fs.existsSync(localFilePath)) {
+      fs.unlinkSync(localFilePath);
+    }
+
     return response;
   } catch (error) {
-    fs.unlinkSync(localFilePath); // remove the locally saved temporary file as the upload operation got failed
+    console.error(" Cloudinary upload failed:");
+    console.error(error);
+
+    if (fs.existsSync(localFilePath)) {
+      fs.unlinkSync(localFilePath);
+    }
+
     return null;
   }
 };
-
-// cloudinary.uploader
-//   .upload("my_image.jpg")
-//   .then((result) => console.log(result));
 
 export { uploadOnCloudinary };
